docker/python/requirements.txt
astropy
pillow
mysql-connector-python
watchdog
xxhash
pyephem
numpy